Move all struct and typedefs to separate header
According to C standard :
ISO/IEC 9899:1999 + TC3 6.2.6/6:
```
The following identifiers have no linkage: an identifier declared to be anything other than an object or a function [...]
```
ISO/IEC 9899:1999 + TC3 6.2.6/6:
```
If an identifier has no linkage, there shall be no more than one declaration of the identifier (in a declarator or type specifier) with the same scope and in the same name space, except for tags as specified in 6.7.2.3.
```

Typedefs cannot be defined multiple times.
Because of that bson_c_lib failes to compile with gcc4.4.2